move scale to base object and fix #2135
